+++ apps/admin-ui-backup/src/services/dump-global-config.ts
@@ -6,13 +6,25 @@
 
 import { postSoapFetchRequest } from '@zextras/ui-shared';
 
-export const dumpGlobalConfig = async (): Promise<any> => {
-  const request: any = {
+import type { DumpGlobalConfigResponse } from '../../types';
+
+type DumpGlobalConfigRequest = {
+  _jsns: string;
+  module: string;
+  action: string;
+};
+
+export const dumpGlobalConfig = async (): Promise<DumpGlobalConfigResponse> => {
+  const request: DumpGlobalConfigRequest = {
     _jsns: 'urn:zimbraAdmin',
     module: 'ZxConfig',
     action: 'dump_global_config',
     // targetServers: serverName
   };
 
-  return postSoapFetchRequest(`/service/admin/soap/zextras`, request, 'zextras');
+  return postSoapFetchRequest<DumpGlobalConfigRequest, DumpGlobalConfigResponse>(
+    `/service/admin/soap/zextras`,
+    request,
+    'zextras',
+  );
 };